Split unknown_064830.c into unknown_065550.c and moved the rest of the rodata
