+++ asm/non_matchings/unknown_064830/alFxParamHdl.s
@@ -1,3 +1,12 @@
+.late_rodata
+glabel jpt_800E6E70
+.word L80063FE8, L8006401C, L8006407C, L80064050, L800640A8, L800640D4, L80064144, L800641C4
+glabel D_800E6E90
+.float 8.320827
+glabel D_800E6E94
+.float 0.04680453
+
+.text
 glabel alFxParamHdl
 /* 064BAC 80063FAC 27BDFFE8 */  addiu $sp, $sp, -0x18
 /* 064BB0 80063FB0 24A3FFFE */  addiu $v1, $a1, -2
@@ -8,12 +17,13 @@ glabel alFxParamHdl
 /* 064BC4 80063FC4 10200093 */  beqz  $at, .L80064214
 /* 064BC8 80063FC8 00803825 */   move  $a3, $a0
 /* 064BCC 80063FCC 000E7080 */  sll   $t6, $t6, 2
-/* 064BD0 80063FD0 3C01800E */  lui   $at, %hi(D_800E6E70) # $at, 0x800e
+/* 064BD0 80063FD0 3C01800E */  lui   $at, %hi(jpt_800E6E70) # $at, 0x800e
 /* 064BD4 80063FD4 002E0821 */  addu  $at, $at, $t6
-/* 064BD8 80063FD8 8C2E6E70 */  lw    $t6, %lo(D_800E6E70)($at)
+/* 064BD8 80063FD8 8C2E6E70 */  lw    $t6, %lo(jpt_800E6E70)($at)
 /* 064BDC 80063FDC 00000000 */  nop   
 /* 064BE0 80063FE0 01C00008 */  jr    $t6
 /* 064BE4 80063FE4 00000000 */   nop   
+glabel L80063FE8
 /* 064BE8 80063FE8 2401FFF8 */  li    $at, -8
 /* 064BEC 80063FEC 01017824 */  and   $t7, $t0, $at
 /* 064BF0 80063FF0 8CF80020 */  lw    $t8, 0x20($a3)
@@ -28,6 +38,7 @@ glabel alFxParamHdl
 /* 064C10 80064010 03095021 */  addu  $t2, $t8, $t1
 /* 064C14 80064014 1000007F */  b     .L80064214
 /* 064C18 80064018 AD4F0000 */   sw    $t7, ($t2)
+glabel L8006401C
 /* 064C1C 8006401C 2401FFF8 */  li    $at, -8
 /* 064C20 80064020 01015824 */  and   $t3, $t0, $at
 /* 064C24 80064024 8CEC0020 */  lw    $t4, 0x20($a3)
@@ -42,6 +53,7 @@ glabel alFxParamHdl
 /* 064C44 80064044 018EC821 */  addu  $t9, $t4, $t6
 /* 064C48 80064048 10000072 */  b     .L80064214
 /* 064C4C 8006404C AF2B0004 */   sw    $t3, 4($t9)
+glabel L80064050
 /* 064C50 80064050 8CF80020 */  lw    $t8, 0x20($a3)
 /* 064C54 80064054 04610003 */  bgez  $v1, .L80064064
 /* 064C58 80064058 000348C3 */   sra   $t1, $v1, 3
@@ -54,6 +66,7 @@ glabel alFxParamHdl
 /* 064C70 80064070 030F5021 */  addu  $t2, $t8, $t7
 /* 064C74 80064074 10000067 */  b     .L80064214
 /* 064C78 80064078 A5480008 */   sh    $t0, 8($t2)
+glabel L8006407C
 /* 064C7C 8006407C 8CED0020 */  lw    $t5, 0x20($a3)
 /* 064C80 80064080 04610003 */  bgez  $v1, .L80064090
 /* 064C84 80064084 000360C3 */   sra   $t4, $v1, 3
@@ -66,6 +79,7 @@ glabel alFxParamHdl
 /* 064C9C 8006409C 01AE5821 */  addu  $t3, $t5, $t6
 /* 064CA0 800640A0 1000005C */  b     .L80064214
 /* 064CA4 800640A4 A568000A */   sh    $t0, 0xa($t3)
+glabel L800640A8
 /* 064CA8 800640A8 8CF90020 */  lw    $t9, 0x20($a3)
 /* 064CAC 800640AC 04610003 */  bgez  $v1, .L800640BC
 /* 064CB0 800640B0 000348C3 */   sra   $t1, $v1, 3
@@ -78,6 +92,7 @@ glabel alFxParamHdl
 /* 064CC8 800640C8 03387821 */  addu  $t7, $t9, $t8
 /* 064CCC 800640CC 10000051 */  b     .L80064214
 /* 064CD0 800640D0 A5E8000C */   sh    $t0, 0xc($t7)
+glabel L800640D4
 /* 064CD4 800640D4 44882000 */  mtc1  $t0, $f4
 /* 064CD8 800640D8 3C01447A */  li    $at, 0x447A0000 # 1000.000000
 /* 064CDC 800640DC 468021A0 */  cvt.s.w $f6, $f4
@@ -107,6 +122,7 @@ glabel alFxParamHdl
 /* 064D38 80064138 01AB4821 */  addu  $t1, $t5, $t3
 /* 064D3C 8006413C 10000035 */  b     .L80064214
 /* 064D40 80064140 E5280010 */   swc1  $f8, 0x10($t1)
+glabel L80064144
 /* 064D44 80064144 8CF90020 */  lw    $t9, 0x20($a3)
 /* 064D48 80064148 04610003 */  bgez  $v1, .L80064158
 /* 064D4C 8006414C 0003C0C3 */   sra   $t8, $v1, 3
@@ -141,6 +157,7 @@ glabel alFxParamHdl
 /* 064DB8 800641B8 462021A0 */  cvt.s.d $f6, $f4
 /* 064DBC 800641BC 10000015 */  b     .L80064214
 /* 064DC0 800641C0 E446001C */   swc1  $f6, 0x1c($v0)
+glabel L800641C4
 /* 064DC4 800641C4 04610003 */  bgez  $v1, .L800641D4
 /* 064DC8 800641C8 000310C3 */   sra   $v0, $v1, 3
 /* 064DCC 800641CC 24610007 */  addiu $at, $v1, 7
